+++ changelog.md
@@ -1283,3 +1283,4 @@ Editor
 Editor
 
 - Added the 'No Camera Easing' property to Portals, which, if true, prevents the camera from easing when you teleport using a Portal.
+- Added moveCamera() and refocusCamera() to the Lua API. See the documentation for more information.